rag_service/pyproject.toml
[project]
name = "rag-service"
version = "1.0.0"
description = "RAG Service - 本地RAG服务，包含完整的LangChain/LangGraph逻辑"
requires-python = ">=3.10"
dependencies = [
    # LangChain/LangGraph依赖
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-anthropic>=1.0.0",
    "anthropic>=0.69.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pinecone>=6.0.0",
    "langchain-pinecone>=0.2.13",
    "jieba>=0.42.1",
    "rank-bm25>=0.2.2",
    # 模型加载依赖
    "langchain-huggingface>=0.1.0",  # HuggingFaceEmbeddings
    "sentence-transformers>=2.2.0",  # CrossEncoderReranker
    "modelscope>=1.9.0",  # ModelScope SDK
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    # 数据库和存储
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.9",
    # 其他必要依赖
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    "pypdf>=6.2.0",  # PDF解析
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []


